# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RidsHealth {
  status: string;
  models_loaded: boolean;
  features_count: number;
  storage_configured: boolean;
  student: string;
  project: string;
  timestamp: string;
}

export interface FeatureImportance { [key: string]: unknown }

export interface ModelInfo {
  model_type: string;
  features_count: number;
  features_list: string[];
  /** @nullable */
  feature_importance: FeatureImportance[] | null;
  status: string;
  student: string;
  project: string;
}

export interface FeatureMap {[key: string]: number}

export interface PredictInput {
  sample?: FeatureMap;
  samples?: FeatureMap[];
  explain?: boolean;
}

export interface PredictionResult {
  predictions: number[];
  probabilities: number[];
  model_used: string;
}

export type PredictionResponseModelInfo = { [key: string]: unknown };

export interface Explanation { [key: string]: unknown }

export interface PredictionResponse {
  status: string;
  timestamp: string;
  student: string;
  project: string;
  samples_count: number;
  results: PredictionResult;
  /** @nullable */
  explanations: Explanation[] | null;
  model_info: PredictionResponseModelInfo;
}

export interface BatchPredictInput {
  samples: FeatureMap[];
}

export interface BatchPredictionItem { [key: string]: unknown }

export interface BatchPredictionResponse {
  status: string;
  timestamp: string;
  student: string;
  project: string;
  total_samples: number;
  results: BatchPredictionItem[];
}

export interface FileUploadInput {
  file: Blob;
}

export interface FilePredictionResponse { [key: string]: unknown }

export interface HistoryItem { [key: string]: unknown }

export interface HistoryResponse {
  status: string;
  history: HistoryItem[];
  count: number;
}