# lib/api-zod/src/generated/types/batchPredictInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FeatureMap } from './featureMap';

export interface BatchPredictInput {
  samples: FeatureMap[];
}