# lib/api-zod/src/generated/types/batchPredictionResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BatchPredictionItem } from './batchPredictionItem';

export interface BatchPredictionResponse {
  status: string;
  timestamp: string;
  student: string;
  project: string;
  total_samples: number;
  results: BatchPredictionItem[];
}